requirements.txt
aws-cdk-lib
boto3
constructs